feat(JWT-3707): move old jwt-tutorial-spring-boot-001 to jwt-tutorial-docker
