+++ jwt-tutorial-docker/pom.xml
@@ -4,8 +4,8 @@
     xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>jwt-tutorial-spring-boot-001</artifactId>
-  <name>jwt-tutorial-spring-boot-001</name>
+  <artifactId>jwt-tutorial-docker</artifactId>
+  <name>jwt-tutorial-docker</name>
   <url>https://levigo.de/info/x/4pZdBg</url>
 
   <parent>
