+++ jwt-tutorial-docker/src/main/.gitignore
