+++ jwt-tutorial-docker/src/main/assembly/all.xml
