+++ jwt-tutorial-docker/src/main/java/org/jwt/Application.gwt.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.8.0//EN" "http://www.gwtproject.org/doctype/2.8.2/gwt-module.dtd">
+<module rename-to="imageviewer">
+	<!-- GWT -->
+    <inherits name="com.google.gwt.user.User"/>
+
+    <!-- jadice web toolkit -->
+    <inherits name="com.levigo.jadice.web.JadiceWebtoolkit"/>
+    <inherits name="com.levigo.jadice.web.Transport"/>
+
+	<!-- The locale configuration with en as default. -->
+	<extend-property name="locale" values="de,en" />
+	<set-property name="locale" value="de,en" />
+	<set-property-fallback name="locale" value="en" />
+
+	<!--  -->
+	<set-configuration-property name="CssResource.style" value="pretty"/>
+    <set-configuration-property name="CssResource.obfuscationPrefix" value="iv"/>
+    <set-configuration-property name="CssResource.enableGss" value="true"/>
+    <set-configuration-property name="CssResource.conversionMode" value="strict"/>
+
+	<!-- Define all subpackages which sources are used in the client. In our case: org.jwt.client -->
+	<source path="client" />
+	<source path="shared" />
+
+	<!-- The entry-point of the application -->
+	<entry-point class="org.jwt.client.ApplicationEntryPoint" />
+</module>
