+++ jwt-tutorial-docker/src/main/java/org/jwt/server/MySpringBootApplication.java
@@ -1,0 +1,21 @@
+package org.jwt.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+import com.levigo.jadice.web.server.spring.autoconfig.EnableJWTSpringBootApplication;
+
+@SpringBootApplication
+@EnableJWTSpringBootApplication("org.jwt.Application")
+public class MySpringBootApplication extends SpringBootServletInitializer {
+	public static void main(String[] args) {
+		SpringApplication.run(MySpringBootApplication.class, args);
+	}
+	
+	@Override
+  protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
+    return super.configure(builder);
+  }
+}
