+++ jwt-tutorial-docker/src/main/webapp/.gitignore
@@ -1,0 +1,2 @@
+WEB-INF
+imageviewer
