+++ jwt-tutorial-docker/src/main/webapp/index.html
