--- jwt-tutorial-spring-boot-001/src/main/webapp/.gitignore
@@ -1,2 +1,0 @@
-WEB-INF
-imageviewer
