+++ pom.xml
@@ -1,23 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.jwt</groupId>
-	<artifactId>jwt-tutorial</artifactId>
-	<version>5.10.52.2</version>
-	<name>jwt-tutorial</name>
-	<description>Getting Started with Jadice Web Toolkit</description>
-	<url>https://levigo.de/info/x/54CaBQ</url>
-	<packaging>pom</packaging>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.jwt</groupId>
+  <artifactId>jwt-tutorial</artifactId>
+  <version>5.10.52.2</version>
+  <name>jwt-tutorial</name>
+  <description>Getting Started with Jadice Web Toolkit</description>
+  <url>https://levigo.de/info/x/54CaBQ</url>
+  <packaging>pom</packaging>
 
-	<modules>
-		<module>jwt-tutorial-000</module>
-		<module>jwt-tutorial-001</module>
-		<module>jwt-tutorial-002</module>
-		<module>jwt-tutorial-003</module>
-		<module>jwt-tutorial-004</module>
-		<module>jwt-tutorial-005</module>
-		<module>jwt-tutorial-performancetest</module>
-		<module>jwt-tutorial-spring-boot-001</module>
+  <modules>
+    <module>jwt-tutorial-000</module>
+    <module>jwt-tutorial-001</module>
+    <module>jwt-tutorial-002</module>
+    <module>jwt-tutorial-003</module>
+    <module>jwt-tutorial-004</module>
+    <module>jwt-tutorial-005</module>
+    <module>jwt-tutorial-docker</module>
+    <module>jwt-tutorial-performancetest</module>
   </modules>
 </project>
